@@ -3,7 +3,7 @@ description: Demonstrates how to use the flutter_blue plugin.
 publish_to: "none"
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -14,9 +14,6 @@ dev_dependencies:
     sdk: flutter
   flutter_blue:
     path: ../
-  e2e: ^0.2.1
-  flutter_driver:
-    sdk: flutter
 
 flutter:
   uses-material-design: true
