@@ -6,10 +6,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 
 class ScanResultTile extends StatelessWidget {
-  const ScanResultTile({Key key, this.result, this.onTap}) : super(key: key);
+  const ScanResultTile({Key? key, required this.result, this.onTap})
+      : super(key: key);
 
   final ScanResult result;
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   Widget _buildTitle(BuildContext context) {
     if (result.device.name.length > 0) {
@@ -48,7 +49,7 @@ class ScanResultTile extends StatelessWidget {
               style: Theme.of(context)
                   .textTheme
                   .caption
-                  .apply(color: Colors.black),
+                  ?.apply(color: Colors.black),
               softWrap: true,
             ),
           ),
@@ -64,7 +65,7 @@ class ScanResultTile extends StatelessWidget {
 
   String getNiceManufacturerData(Map<int, List<int>> data) {
     if (data.isEmpty) {
-      return null;
+      return 'N/A';
     }
     List<String> res = [];
     data.forEach((id, bytes) {
@@ -76,7 +77,7 @@ class ScanResultTile extends StatelessWidget {
 
   String getNiceServiceData(Map<String, List<int>> data) {
     if (data.isEmpty) {
-      return null;
+      return 'N/A';
     }
     List<String> res = [];
     data.forEach((id, bytes) {
@@ -101,12 +102,8 @@ class ScanResultTile extends StatelessWidget {
             context, 'Complete Local Name', result.advertisementData.localName),
         _buildAdvRow(context, 'Tx Power Level',
             '${result.advertisementData.txPowerLevel ?? 'N/A'}'),
-        _buildAdvRow(
-            context,
-            'Manufacturer Data',
-            getNiceManufacturerData(
-                    result.advertisementData.manufacturerData) ??
-                'N/A'),
+        _buildAdvRow(context, 'Manufacturer Data',
+            getNiceManufacturerData(result.advertisementData.manufacturerData)),
         _buildAdvRow(
             context,
             'Service UUIDs',
@@ -114,7 +111,7 @@ class ScanResultTile extends StatelessWidget {
                 ? result.advertisementData.serviceUuids.join(', ').toUpperCase()
                 : 'N/A'),
         _buildAdvRow(context, 'Service Data',
-            getNiceServiceData(result.advertisementData.serviceData) ?? 'N/A'),
+            getNiceServiceData(result.advertisementData.serviceData)),
       ],
     );
   }
@@ -124,7 +121,8 @@ class ServiceTile extends StatelessWidget {
   final BluetoothService service;
   final List<CharacteristicTile> characteristicTiles;
 
-  const ServiceTile({Key key, this.service, this.characteristicTiles})
+  const ServiceTile(
+      {Key? key, required this.service, required this.characteristicTiles})
       : super(key: key);
 
   @override
@@ -137,10 +135,8 @@ class ServiceTile extends StatelessWidget {
           children: <Widget>[
             Text('Service'),
             Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}',
-                style: Theme.of(context)
-                    .textTheme
-                    .body1
-                    .copyWith(color: Theme.of(context).textTheme.caption.color))
+                style: Theme.of(context).textTheme.body1?.copyWith(
+                    color: Theme.of(context).textTheme.caption?.color))
           ],
         ),
         children: characteristicTiles,
@@ -158,14 +154,14 @@ class ServiceTile extends StatelessWidget {
 class CharacteristicTile extends StatelessWidget {
   final BluetoothCharacteristic characteristic;
   final List<DescriptorTile> descriptorTiles;
-  final VoidCallback onReadPressed;
-  final VoidCallback onWritePressed;
-  final VoidCallback onNotificationPressed;
+  final VoidCallback? onReadPressed;
+  final VoidCallback? onWritePressed;
+  final VoidCallback? onNotificationPressed;
 
   const CharacteristicTile(
-      {Key key,
-      this.characteristic,
-      this.descriptorTiles,
+      {Key? key,
+      required this.characteristic,
+      required this.descriptorTiles,
       this.onReadPressed,
       this.onWritePressed,
       this.onNotificationPressed})
@@ -187,8 +183,8 @@ class CharacteristicTile extends StatelessWidget {
                 Text('Characteristic'),
                 Text(
                     '0x${characteristic.uuid.toString().toUpperCase().substring(4, 8)}',
-                    style: Theme.of(context).textTheme.body1.copyWith(
-                        color: Theme.of(context).textTheme.caption.color))
+                    style: Theme.of(context).textTheme.body1?.copyWith(
+                        color: Theme.of(context).textTheme.caption?.color))
               ],
             ),
             subtitle: Text(value.toString()),
@@ -200,13 +196,13 @@ class CharacteristicTile extends StatelessWidget {
               IconButton(
                 icon: Icon(
                   Icons.file_download,
-                  color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+                  color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
                 ),
                 onPressed: onReadPressed,
               ),
               IconButton(
                 icon: Icon(Icons.file_upload,
-                    color: Theme.of(context).iconTheme.color.withOpacity(0.5)),
+                    color: Theme.of(context).iconTheme.color?.withOpacity(0.5)),
                 onPressed: onWritePressed,
               ),
               IconButton(
@@ -214,7 +210,7 @@ class CharacteristicTile extends StatelessWidget {
                     characteristic.isNotifying
                         ? Icons.sync_disabled
                         : Icons.sync,
-                    color: Theme.of(context).iconTheme.color.withOpacity(0.5)),
+                    color: Theme.of(context).iconTheme.color?.withOpacity(0.5)),
                 onPressed: onNotificationPressed,
               )
             ],
@@ -228,11 +224,14 @@ class CharacteristicTile extends StatelessWidget {
 
 class DescriptorTile extends StatelessWidget {
   final BluetoothDescriptor descriptor;
-  final VoidCallback onReadPressed;
-  final VoidCallback onWritePressed;
+  final VoidCallback? onReadPressed;
+  final VoidCallback? onWritePressed;
 
   const DescriptorTile(
-      {Key key, this.descriptor, this.onReadPressed, this.onWritePressed})
+      {Key? key,
+      required this.descriptor,
+      this.onReadPressed,
+      this.onWritePressed})
       : super(key: key);
 
   @override
@@ -247,7 +246,7 @@ class DescriptorTile extends StatelessWidget {
               style: Theme.of(context)
                   .textTheme
                   .body1
-                  .copyWith(color: Theme.of(context).textTheme.caption.color))
+                  ?.copyWith(color: Theme.of(context).textTheme.caption?.color))
         ],
       ),
       subtitle: StreamBuilder<List<int>>(
@@ -261,14 +260,14 @@ class DescriptorTile extends StatelessWidget {
           IconButton(
             icon: Icon(
               Icons.file_download,
-              color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+              color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
             ),
             onPressed: onReadPressed,
           ),
           IconButton(
             icon: Icon(
               Icons.file_upload,
-              color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+              color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
             ),
             onPressed: onWritePressed,
           )
@@ -279,7 +278,7 @@ class DescriptorTile extends StatelessWidget {
 }
 
 class AdapterStateTile extends StatelessWidget {
-  const AdapterStateTile({Key key, @required this.state}) : super(key: key);
+  const AdapterStateTile({Key? key, required this.state}) : super(key: key);
 
   final BluetoothState state;
 
@@ -294,7 +293,7 @@ class AdapterStateTile extends StatelessWidget {
         ),
         trailing: Icon(
           Icons.error,
-          color: Theme.of(context).primaryTextTheme.subhead.color,
+          color: Theme.of(context).primaryTextTheme.subhead?.color,
         ),
       ),
     );
