@@ -14,9 +14,8 @@ class FlutterBlue {
 
   /// Singleton boilerplate
   FlutterBlue._() {
-    _channel.setMethodCallHandler((MethodCall call) {
+    _channel.setMethodCallHandler((MethodCall call) async {
       _methodStreamController.add(call);
-      return;
     });
 
     _setLogLevelIfAvailable();
@@ -38,10 +37,40 @@ class FlutterBlue {
   /// Checks if Bluetooth functionality is turned on
   Future<bool> get isOn => _channel.invokeMethod('isOn').then<bool>((d) => d);
 
+  ///Tries to turn on Bluetooth,
+  ///
+  ///returns true if bluetooth is being turned on.
+  ///You have to listen for a stateChange to ON to ensure bluetooth is already running
+  ///
+  ///returns false if an error occured or bluetooth is already running
+  ///
+  Future<bool> turnOn() {
+    return _channel.invokeMethod('turnOn').then<bool>((d) => d);
+  }
+
+  ///Tries to turn off Bluetooth,
+  ///
+  ///returns true if bluetooth is being turned off.
+  ///You have to listen for a stateChange to OFF to ensure bluetooth is turned off
+  ///
+  ///returns false if an error occured
+  ///
+  Future<bool> turnOff() {
+    return _channel.invokeMethod('turnOff').then<bool>((d) => d);
+  }
+
   BehaviorSubject<bool> _isScanning = BehaviorSubject.seeded(false);
   Stream<bool> get isScanning => _isScanning.stream;
 
   BehaviorSubject<List<ScanResult>> _scanResults = BehaviorSubject.seeded([]);
+
+  /// Returns a stream that is a list of [ScanResult] results while a scan is in progress.
+  ///
+  /// The list emitted is all the scanned results as of the last initiated scan. When a scan is
+  /// first started, an empty list is emitted. The returned stream is never closed.
+  ///
+  /// One use for [scanResults] is as the stream in a StreamBuilder to display the
+  /// results of a scan in real time while the scan is in progress.
   Stream<List<ScanResult>> get scanResults => _scanResults.stream;
 
   PublishSubject _stopScanPill = new PublishSubject();
@@ -75,13 +104,17 @@ class FlutterBlue {
     }
   }
 
-  /// Starts a scan for Bluetooth Low Energy devices
-  /// Timeout closes the stream after a specified [Duration]
+  /// Starts a scan for Bluetooth Low Energy devices and returns a stream
+  /// of the [ScanResult] results as they are received.
+  ///
+  /// timeout calls stopStream after a specified [Duration].
+  /// You can also get a list of ongoing results in the [scanResults] stream.
+  /// If scanning is already in progress, this will throw an [Exception].
   Stream<ScanResult> scan({
     ScanMode scanMode = ScanMode.lowLatency,
     List<Guid> withServices = const [],
     List<Guid> withDevices = const [],
-    Duration timeout,
+    Duration? timeout,
     bool allowDuplicates = false,
   }) async* {
     var settings = protos.ScanSettings.create()
@@ -122,7 +155,7 @@ class FlutterBlue {
         .map((buffer) => new protos.ScanResult.fromBuffer(buffer))
         .map((p) {
       final result = new ScanResult.fromProto(p);
-      final list = _scanResults.value;
+      final list = _scanResults.value ?? [];
       int index = list.indexOf(result);
       if (index != -1) {
         list[index] = result;
@@ -134,11 +167,19 @@ class FlutterBlue {
     });
   }
 
+  /// Starts a scan and returns a future that will complete once the scan has finished.
+  ///
+  /// Once a scan is started, call [stopScan] to stop the scan and complete the returned future.
+  ///
+  /// timeout automatically stops the scan after a specified [Duration].
+  ///
+  /// To observe the results while the scan is in progress, listen to the [scanResults] stream,
+  /// or call [scan] instead.
   Future startScan({
     ScanMode scanMode = ScanMode.lowLatency,
     List<Guid> withServices = const [],
     List<Guid> withDevices = const [],
-    Duration timeout,
+    Duration? timeout,
     bool allowDuplicates = false,
   }) async {
     await scan(
@@ -229,8 +270,6 @@ class DeviceIdentifier {
 }
 
 class ScanResult {
-  const ScanResult({this.device, this.advertisementData, this.rssi});
-
   ScanResult.fromProto(protos.ScanResult p)
       : device = new BluetoothDevice.fromProto(p.device),
         advertisementData =
@@ -259,19 +298,11 @@ class ScanResult {
 
 class AdvertisementData {
   final String localName;
-  final int txPowerLevel;
+  final int? txPowerLevel;
   final bool connectable;
   final Map<int, List<int>> manufacturerData;
   final Map<String, List<int>> serviceData;
   final List<String> serviceUuids;
-
-  AdvertisementData(
-      {this.localName,
-      this.txPowerLevel,
-      this.connectable,
-      this.manufacturerData,
-      this.serviceData,
-      this.serviceUuids});
 
   AdvertisementData.fromProto(protos.AdvertisementData p)
       : localName = p.localName,
