@@ -2,7 +2,11 @@
 // All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-#import <Flutter/Flutter.h>
+#if TARGET_OS_OSX
+ #import <FlutterMacOS/FlutterMacOS.h>
+#else
+ #import <Flutter/Flutter.h>
+#endif
 #import <CoreBluetooth/CoreBluetooth.h>
 
 #define NAMESPACE @"plugins.pauldemarco.com/flutter_blue"
