@@ -15,7 +15,7 @@ class BluetoothDescriptor {
   BehaviorSubject<List<int>> _value;
   Stream<List<int>> get value => _value.stream;
 
-  List<int> get lastValue => _value.value;
+  List<int> get lastValue => _value.value ?? [];
 
   BluetoothDescriptor.fromProto(protos.BluetoothDescriptor p)
       : uuid = new Guid(p.uuid),
@@ -84,6 +84,6 @@ class BluetoothDescriptor {
 
   @override
   String toString() {
-    return 'BluetoothDescriptor{uuid: $uuid, deviceId: $deviceId, serviceUuid: $serviceUuid, characteristicUuid: $characteristicUuid, value: ${_value?.value}}';
+    return 'BluetoothDescriptor{uuid: $uuid, deviceId: $deviceId, serviceUuid: $serviceUuid, characteristicUuid: $characteristicUuid, value: ${_value.value}}';
   }
 }
